requirements.txt
Django==5.2
cloudinary==1.44.0
Pillow==11.2.1
mysqlclient=2.2.7
